apai_demo/requirements.txt
google-generativeai
matplotlib
scipy
numpy
